requirements.txt
PyQt6
opencv-python
Pillow
SpeechRecognition
google-generativeai
pyaudio
tensorflow
